Add HANDLE-QVM-MEMORY-ESTIMATE RPC handler in app-ng
The number returned represents the number of bytes required to store
the QVM state, i.e. to store the amplitudes of the wavefunction for a
PURE-STATE simulation or for the density matrix for a
FULL-DENSITY-MATRIX simulation. Note that the memory required to store
QVM state represents a lower-bound on the memory required for
simulation, since in the course of simulating programs the QVM will
allocate additional memory. For the most part, these additional
allocations are small (compared to the QVM state), short-lived, and
difficult to predict in advance. We also ignore any memory allocated
for the classical memory subsystem of the QVM, which is bounded by
QVM::**CLASSICAL-MEMORY-SIZE-LIMIT** (64K by default).

Closes #190